+++ webapp/src/components/backstage/playbook_runs/playbook_run/finish_run.tsx
@@ -4,13 +4,47 @@
 import React from 'react';
 import styled from 'styled-components';
 import {useIntl} from 'react-intl';
+import {useDispatch} from 'react-redux';
 import Icon from '@mdi/react';
 import {mdiFlagOutline} from '@mdi/js';
 
 import {PlaybookRun, PlaybookRunStatus} from 'src/types/playbook_run';
 import {TertiaryButton} from 'src/components/assets/buttons';
+import {finishRun} from 'src/client';
+import {modals} from 'src/webapp_globals';
+import {outstandingTasks} from 'src/components/modals/update_run_status_modal';
+import {makeUncontrolledConfirmModalDefinition} from 'src/components/widgets/confirmation_modal';
 
-import withFinishRun from './hoc_finish_run';
+export const useOnFinishRun = (playbookRun: PlaybookRun) => {
+    const dispatch = useDispatch();
+    const {formatMessage} = useIntl();
+
+    return () => {
+        const outstanding = outstandingTasks(playbookRun.checklists);
+        let confirmationMessage = formatMessage({defaultMessage: 'Are you sure you want to finish the run?'});
+        if (outstanding > 0) {
+            confirmationMessage = formatMessage(
+                {defaultMessage: 'There {outstanding, plural, =1 {is # outstanding task} other {are # outstanding tasks}}. Are you sure you want to finish the run?'},
+                {outstanding}
+            );
+        }
+
+        const onConfirm = () => {
+            finishRun(playbookRun.id);
+        };
+
+        dispatch(modals.openModal(makeUncontrolledConfirmModalDefinition({
+            show: true,
+            title: formatMessage({defaultMessage: 'Confirm finish run'}),
+            message: confirmationMessage,
+            confirmButtonText: formatMessage({defaultMessage: 'Finish run'}),
+            onConfirm,
+            // eslint-disable-next-line @typescript-eslint/no-empty-function
+            onCancel: () => {},
+        })));
+    };
+};
+
 interface Props {
     playbookRun: PlaybookRun;
 }
@@ -18,11 +52,11 @@ interface Props {
 const FinishRun = ({playbookRun}: Props) => {
     const {formatMessage} = useIntl();
 
+    const onFinishRun = useOnFinishRun(playbookRun);
+
     if (playbookRun.current_status === PlaybookRunStatus.Finished) {
         return null;
     }
-
-    const Finish = withFinishRun(FinishRunButton);
 
     return (
         <Container>
@@ -35,9 +69,9 @@ const FinishRun = ({playbookRun}: Props) => {
                 </IconWrapper>
                 <Text>{formatMessage({defaultMessage: 'Time to wrap up?'})}</Text>
                 <RightWrapper>
-                    <Finish playbookRun={playbookRun}>
+                    <FinishRunButton onClick={onFinishRun}>
                         {formatMessage({defaultMessage: 'Finish run'})}
-                    </Finish>
+                    </FinishRunButton>
                 </RightWrapper>
             </Content>
         </Container>
